+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+Version 0.0.4 :
+
+Feature:
+ - On search the text preview is now centered on the string searched.
+Fix:
+ - The search included previously the URL text, but the results where incoherent. The search is now only on title and fact text.
+
 Version 0.0.3 :
 
 Feature: